added user confirmation to run program& added comments

+++ scrape-from-csv/scrape_and_create_db.py
@@ -3,28 +3,23 @@ from urllib2 import urlopen
 from bs4 import BeautifulSoup
 import pymongo
 import sys
+from settings import MONGODB_HOST, MONGODB_PORT, DBS_NAME, COLLECTION_NAME, CSV_FILE_NAME
 
 '''
-Creates initial MongoDb database with collection name
+Creates a MongoDb database with collection name
 scrape for data, connecto to MongoDB, create db, create, collection, and
 insert to collection.
 '''
 
-MONGODB_HOST = 'localhost'
-MONGODB_PORT = 27017
-DBS_NAME = 'test'
-COLLECTION_NAME = 'test_collection'
-CSV_FILE_NAME = 'urls.csv'
-
-# open csv file and return the row
+# open csv file and return the row in a list
 def readCsvFile(filename):
     try:
         with open(filename) as csvDataFile:
             csvReader = csv.reader(csvDataFile)
             for row in csvReader:
                 return row
-    except IOError as e:
-        print 'Error reading file: %s found' % e
+    except IOError as error:
+        print 'Error reading file: %s' % error
         print 'exiting program...'
         sys.exit()
 
@@ -34,17 +29,16 @@ def mongo_connect():
         conn = pymongo.MongoClient(MONGODB_HOST, MONGODB_PORT)
         print "Mongo is connected!"
         return conn
-    except pymongo.errors.ConnectionFailure as e:
-        print "Could not connect to MongoDB: %s" % e
+    except pymongo.errors.ConnectionFailure as error:
+        print "Could not connect to MongoDB: %s" % error
 
 # Parse html of webpage and find title and description and store in a dictionary
-def get_data(urls):
+def scrape_data(urls):
     data_array = []
     for url in urls:
-        html_text = urlopen(url).read()
+        html_text = urlopen(url).read() # how to narrow the read down to only <head> tag? (make more efficient?)
         soup = BeautifulSoup(html_text, "lxml")
         # get title and description from each url
-        # Only need to search through <head>. Any way to narrow the parse of each page to that?
         site_title = soup.title.string
         site_description = soup.html.head.find("meta", attrs={"name":"description"})
         # need to create new dicionary for each new item/iteration to be able to append to list later
@@ -57,13 +51,28 @@ def get_data(urls):
         data_array.append(data)
     return data_array
 
+
+print "This program will scrape title and description of each url included in the csv file,"
+print "and write the results to a MongoDB database. (make sure you're connected to MongoDB)"
+print "\nWARNING: This will create a database named: '%s', and a collection named: '%s'" % (DBS_NAME, COLLECTION_NAME)
+print "WARNING: The collection will be overwritten if one already exists within the same db"
+print "NOTE: edit settings.py to change db and collection names as well as connection settings\n"
+
+user_input = raw_input("Press Enter to proceed... (or type 'quit' or 'exit' to cancel): ")
+
+
+# check user wants to proceed before doing any work
+if user_input == 'quit' or user_input == 'exit':
+    print 'exiting program...'
+    sys.exit()
+
+# get a list of url from a csv file
 my_urls = readCsvFile(CSV_FILE_NAME)
-
-scraped_data = get_data(my_urls)
-
+scraped_data = scrape_data(my_urls)
 # Insert data to MongoDB
 conn = mongo_connect() #connect to MongoDB
 db = conn[DBS_NAME] # db name
 coll = db[COLLECTION_NAME] # db collectinon name
 coll.drop()  # remove the collection to avoid duplicates when testing
 coll.insert(scraped_data) # insert scraped data to db
+
